Implement Add new structure
Fix edit existing structure
Implement Names Panel (merged from branch for task 797)

## Before submitting this PR for code review, please make sure:
- [X] "Describe the changes being made to the code and why" above has been replaced
- [X] All new files have the standard license headers
- [X] CodeMaid has been run against all new or updated files
- [X] All NuGet packages have been updated and file /README.md has been updated
- [X] The code builds clean (please check Debug and Setup configurations) without any errors or new warnings
- [X] All Unit Tests pass
- [X] You have followed the [Chem4Word Coding Standards](https://chem4word.visualstudio.com/C4W-VNext/_wiki/wikis/C4W-VNext.wiki/2/Coding-Standards)
- [X] You have updated the change history (/src/Chem4Word.V3/Data/Chem4Word-Versions.xml)
- [X] If appropriate you have updated the manual (/docs/Chem4Word-Version3-3-User-Manual.docx)
- [X] The build pipeline is green

Related work items: #797, #1029

+++ src/PlugIns/Drivers/Chem4Word.Driver.Open/SqLite/Library.cs
@@ -242,26 +242,30 @@ namespace Chem4Word.Driver.Open.SqLite
 
                 command.ExecuteNonQuery();
 
-                // Delete old record from Properties
-                sb = new StringBuilder();
+                // Properties table only exists in 3.3.1 or greater
+                if (version >= new Version(3, 3, 1))
+                {
+                    // Delete old record
+                    sb = new StringBuilder();
 
-                sb.AppendLine("DELETE FROM Properties");
-                sb.AppendLine("WHERE Key = 'Version'");
+                    sb.AppendLine("DELETE FROM Properties");
+                    sb.AppendLine("WHERE Key = 'Version'");
 
-                command = new SQLiteCommand(sb.ToString(), conn);
-                command.ExecuteNonQuery();
+                    command = new SQLiteCommand(sb.ToString(), conn);
+                    command.ExecuteNonQuery();
 
-                // Add new record in Properties
-                sb = new StringBuilder();
+                    // Add new record in Properties
+                    sb = new StringBuilder();
 
-                sb.AppendLine("INSERT INTO Properties");
-                sb.AppendLine(" (Key, Value)");
-                sb.AppendLine("VALUES");
-                sb.AppendLine(" ('Version', @version)");
+                    sb.AppendLine("INSERT INTO Properties");
+                    sb.AppendLine(" (Key, Value)");
+                    sb.AppendLine("VALUES");
+                    sb.AppendLine(" ('Version', @version)");
 
-                command = new SQLiteCommand(sb.ToString(), conn);
-                command.Parameters.Add("@version", DbType.String, versionString.Length).Value = versionString;
-                command.ExecuteNonQuery();
+                    command = new SQLiteCommand(sb.ToString(), conn);
+                    command.Parameters.Add("@version", DbType.String, versionString.Length).Value = versionString;
+                    command.ExecuteNonQuery();
+                }
             }
             catch (Exception ex)
             {
